ManyToOneBuffer - more tests, minor fixes and proper setting up of mocks for _buffer.GetAtomic* (out of the box all are set in SetUp as proper mocks are based on AtomicInt, AtomicLong

+++ src/RampUp/Ring/RingBufferDescriptor.cs
@@ -110,7 +110,7 @@ namespace RampUp.Ring
 
         public static long MakeHeader(int length, int messageTypeId)
         {
-            return (messageTypeId << 32) | length;
+            return ((long)messageTypeId << 32) | length;
         }
 
         public static int RecordLength(long header)
